New AbstractLinkFactory and split LinkFactory
